Add timeouts to bulk copy command, update testNegativeTimeOut test
