+++ src/main/java/com/microsoft/sqlserver/jdbc/ErrorConstants.java
@@ -1,0 +1,6 @@
+package com.microsoft.sqlserver.jdbc;
+
+public class ErrorConstants {
+    public static String TIMEOUT_ERROR = "R_queryTimedOut";
+    public static String CONNECTION_CLOSED = "R_connectionIsClosed";
+}
